refactor(editor): move marks runtime into package
