+++ packages/doc-editor/package.json
@@ -6,7 +6,8 @@
   "description": "Proof SDK editor runtime and editor-facing plugins",
   "dependencies": {
     "@proof/core": "0.1.0",
-    "@milkdown/kit": "^7.5.0"
+    "@milkdown/kit": "^7.5.0",
+    "y-prosemirror": "^1.2.19"
   },
   "exports": {
     ".": "./src/index.ts",
